# ClientMS-/src/main/java/com/bootcamp/clientms/dto/request/UpdateClientRequest.java
package com.bootcamp.clientms.dto.request;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class UpdateClientRequest {
  @NotBlank(message = "First name is required")
  private String firstName;
  @NotBlank(message = "Last name is required")
  private String lastName;
  @NotBlank(message = "Email is required")
  @Email(message = "Email must be valid")
  private String email;
}